Fix overflow

+++ test/aoj/DSL/DSL_2_G-RAQRSQ.test.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 int main() {
     int n, q; cin >> n >> q;
-    RangeAddQueryRangeSumQuery<int, int> seg(n);
+    RangeAddQueryRangeSumQuery<ll, ll> seg(n);
     rep (q) {
         int t; cin >> t;
         if (t == 0) {
